requirements.txt
flask
openai
pypdf
python-dotenv
gunicorn